# storage/api-docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Nutrients API",
        "description": "REST API for managing ingredients and their nutritional data.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/ingredients": {
            "get": {
                "tags": [
                    "Ingredients"
                ],
                "summary": "List ingredients (paginated)",
                "operationId": "5d595349bcfde36b13640a043eb2db57",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of ingredients",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Ingredient"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 500
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 25
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": 20
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "frontend": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Ingredients"
                ],
                "summary": "Create a new ingredient",
                "operationId": "0d674a4f6ca04a7604deddd5ffc1d160",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "source",
                                    "name",
                                    "default_amount",
                                    "default_amount_unit_id"
                                ],
                                "properties": {
                                    "source": {
                                        "type": "string",
                                        "example": "USDA"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Broccoli, raw"
                                    },
                                    "external_id": {
                                        "type": "string",
                                        "example": "321360",
                                        "nullable": true
                                    },
                                    "class": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "broccoli-raw",
                                        "nullable": true
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "default_amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100
                                    },
                                    "default_amount_unit_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "brand_id": {
                                        "type": "integer",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ingredient"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "frontend": []
                    }
                ]
            }
        },
        "/api/ingredients/{ingredient}": {
            "get": {
                "tags": [
                    "Ingredients"
                ],
                "summary": "Get a single ingredient with all relationships",
                "operationId": "b5bdb82e230552aabcf49f77306133a5",
                "parameters": [
                    {
                        "name": "ingredient",
                        "in": "path",
                        "description": "Ingredient ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ingredient resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ingredient"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "frontend": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Ingredients"
                ],
                "summary": "Update an ingredient",
                "operationId": "ce3cdee317f0ef1b5d9a3cbc18b101ca",
                "parameters": [
                    {
                        "name": "ingredient",
                        "in": "path",
                        "description": "Ingredient ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "source": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "external_id": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "class": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "slug": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "default_amount": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "default_amount_unit_id": {
                                        "type": "integer"
                                    },
                                    "brand_id": {
                                        "type": "integer",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ingredient"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "frontend": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Ingredients"
                ],
                "summary": "Soft-delete an ingredient",
                "operationId": "b519f81034a5ac4cdff57d149fd05c00",
                "parameters": [
                    {
                        "name": "ingredient",
                        "in": "path",
                        "description": "Ingredient ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "frontend": []
                    }
                ]
            }
        },
        "/api/ingredients/search": {
            "post": {
                "tags": [
                    "Ingredients"
                ],
                "summary": "Full-text search ingredients",
                "operationId": "2fae9b559b45f1504e94f06123bfe98c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "query"
                                ],
                                "properties": {
                                    "query": {
                                        "type": "string",
                                        "example": "broccoli"
                                    },
                                    "page": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Search results",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "hits": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Ingredient"
                                            }
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 3
                                        },
                                        "page": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "frontend": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Brand": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Acme Foods"
                    },
                    "slug": {
                        "type": "string",
                        "example": "acme-foods"
                    },
                    "owner": {
                        "type": "string",
                        "example": "Acme Corp",
                        "nullable": true
                    },
                    "country": {
                        "type": "string",
                        "example": "US",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "IngredientCategory": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Vegetables"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Ingredient": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "external_id": {
                        "type": "string",
                        "example": "321360",
                        "nullable": true
                    },
                    "source": {
                        "type": "string",
                        "example": "USDA",
                        "nullable": true
                    },
                    "class": {
                        "type": "string",
                        "example": "FoodItem",
                        "nullable": true
                    },
                    "name": {
                        "type": "string",
                        "example": "Broccoli, raw"
                    },
                    "slug": {
                        "type": "string",
                        "example": "broccoli-raw"
                    },
                    "description": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "default_amount": {
                        "type": "number",
                        "format": "float",
                        "example": 100,
                        "nullable": true
                    },
                    "sync_status": {
                        "type": "string",
                        "example": "synced"
                    },
                    "brand": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Brand"
                            }
                        ],
                        "nullable": true
                    },
                    "default_amount_unit": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Unit"
                            }
                        ],
                        "nullable": true
                    },
                    "nutrients": {
                        "type": "array",
                        "items": {
                            "allOf": [
                                {
                                    "$ref": "#/components/schemas/Nutrient"
                                },
                                {
                                    "properties": {
                                        "pivot": {
                                            "$ref": "#/components/schemas/IngredientNutrientPivot"
                                        }
                                    },
                                    "type": "object"
                                }
                            ]
                        },
                        "nullable": true
                    },
                    "nutrition_facts": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/IngredientNutritionFact"
                        },
                        "nullable": true
                    },
                    "categories": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/IngredientCategory"
                        },
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Nutrient": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "external_id": {
                        "type": "string",
                        "example": "1004",
                        "nullable": true
                    },
                    "name": {
                        "type": "string",
                        "example": "Vitamin C"
                    },
                    "description": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "slug": {
                        "type": "string",
                        "example": "vitamin-c"
                    },
                    "iu_to_canonical_factor": {
                        "type": "number",
                        "format": "float",
                        "example": null,
                        "nullable": true
                    },
                    "is_label_standard": {
                        "type": "boolean",
                        "example": true
                    },
                    "display_order": {
                        "type": "integer",
                        "example": 10,
                        "nullable": true
                    },
                    "sync_status": {
                        "type": "string",
                        "example": "synced"
                    },
                    "source": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Source"
                            }
                        ],
                        "nullable": true
                    },
                    "canonical_unit": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Unit"
                            }
                        ],
                        "nullable": true
                    },
                    "parent": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Nutrient"
                            }
                        ],
                        "nullable": true
                    },
                    "children": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Nutrient"
                        }
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/NutrientTag"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "Unit": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Gram"
                    },
                    "abbreviation": {
                        "type": "string",
                        "example": "g"
                    },
                    "type": {
                        "type": "string",
                        "example": "mass",
                        "nullable": true
                    },
                    "to_base_factor": {
                        "type": "number",
                        "format": "float",
                        "example": 1,
                        "nullable": true
                    },
                    "base_unit_id": {
                        "type": "integer",
                        "example": null,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "frontend": {
                "type": "http",
                "description": "Bearer token obtained from POST /api/auth/login. Also requires X-Refresh-Token, X-Application-Name, and X-Client-Url headers on every request.",
                "scheme": "bearer"
            }
        }
    },
...
}